Remove redundant translation overrides
Removed several places where a submod would provide its own string, such as object name even though it is completely unnecessary and is identical to string it intends to replace.

This removes need to provide translation override in a submod for a couple of strings.

Adjusted czech translation to account for these changes